docs: elaborate design
