+++ osteoid/formats/ostd/header.py
@@ -1,142 +1,20 @@
-from typing import Optional, Any
+from typing import Optional, Any, Literal
 
-from enum import IntEnum
-import uuid 
+from .types import (
+  CompressionType,
+  DataType,
+  EdgeRepresentation,
+  GraphType,
+  PhysicalUnit,
+  SpaceType,
+  TO_DATATYPE, 
+  FROM_DATATYPE,
+)
+
+import uuid
 
 import numpy as np
 import numpy.typing as npt
-
-class PhysicalUnit(IntEnum):
-  VOXEL = 0
-
-  ANGSTROM = 1
-  FEMTOMETER = 2
-  PICOMETER = 3
-  NANOMETER = 4
-  MICROMETER = 5
-  MILLIMETER = 6
-  CENTIMETER = 7
-  METER = 8
-  KILOMETER = 9
-  MEGAMETER = 10
-  LIGHTYEAR = 11
-  PARSEC = 12
-
-  MIL = 13
-  INCH = 14
-  FOOT = 15
-  YARD = 16
-  STATUTE_MILE = 17
-  NAUTICAL_MILE = 18
-
-TO_UNIT = {
-  "vx": PhysicalUnit.VOXEL,
-  "voxel": PhysicalUnit.VOXEL,
-
-  "A": PhysicalUnit.ANGSTROM,
-  "angstrom": PhysicalUnit.ANGSTROM,
-  
-  "fm": PhysicalUnit.FEMTOMETER,
-  "femtometer": PhysicalUnit.FEMTOMETER,
-  
-  "pm": PhysicalUnit.PICOMETER,
-  "picometer": PhysicalUnit.PICOMETER,
-  
-  "nm": PhysicalUnit.NANOMETER,
-  "nanometer": PhysicalUnit.NANOMETER,
-  
-  "um": PhysicalUnit.MICROMETER,
-  "micrometer": PhysicalUnit.MICROMETER,
-  "micron": PhysicalUnit.MICROMETER,
-  
-  "mm": PhysicalUnit.MILLIMETER,
-  "millimeter": PhysicalUnit.MILLIMETER,
-
-  "cm": PhysicalUnit.CENTIMETER,
-  "centimeter": PhysicalUnit.CENTIMETER,
-  
-  "m": PhysicalUnit.METER,
-  "meter": PhysicalUnit.METER,
-  
-  "km": PhysicalUnit.KILOMETER,
-  "kilometer": PhysicalUnit.KILOMETER,
-
-  "Mm": PhysicalUnit.MEGAMETER,
-  
-  "ly": PhysicalUnit.LIGHTYEAR,
-  "lightyear": PhysicalUnit.LIGHTYEAR,
-
-  "pc": PhysicalUnit.PARSEC,
-  "parsec": PhysicalUnit.PARSEC,
-
-  "mil": PhysicalUnit.MIL,
-  
-  "in": PhysicalUnit.INCH,
-  "inch": PhysicalUnit.INCH,
-  "inches": PhysicalUnit.INCH,
-  
-  "ft": PhysicalUnit.FOOT,
-  "foot": PhysicalUnit.FOOT,
-  "feet": PhysicalUnit.FOOT,
-
-  "yd": PhysicalUnit.YARD,
-  "yard": PhysicalUnit.YARD,
-  
-  "mi": PhysicalUnit.STATUTE_MILE,
-  "mile": PhysicalUnit.STATUTE_MILE,
-  
-  "nmi": PhysicalUnit.NAUTICAL_MILE,
-}
-
-class CompressionType(IntEnum):
-  NONE = 0
-  GZIP = 1
-  BZIP2 = 2
-  ZSTD = 3
-
-class DataType(IntEnum):
-  F8 = 0
-  F16 = 1
-  F32 = 2
-  F64 = 3
-  U8 = 4
-  U16 = 5
-  U32 = 6
-  U64 = 7
-  I8 = 8
-  I16 = 9
-  I32 = 10
-  I64 = 11
-  BOOLEAN = 12
-  PACKED_BOOLEAN = 13
-
-TO_DATATYPE = {
-  np.float16: DataType.F16,
-  np.float32: DataType.F32,
-  np.float64: DataType.F64,
-  np.uint8: DataType.U8,
-  np.uint16: DataType.U16,
-  np.uint32: DataType.U32,
-  np.uint64: DataType.U64,
-  np.int8: DataType.I8,
-  np.int16: DataType.I16,
-  np.int32: DataType.I32,
-  np.int64: DataType.I64,
-}
-
-FROM_DATATYPE = { v,k for k,v in TO_DATATYPE.items() }
-
-class EdgeRepresentation(IntEnum):
-  PAIR = 0
-  PARENT = 1
-
-class GraphType(IntEnum):
-  GRAPH = 0
-  TREE = 1
-
-class SpaceType(IntEnum):
-  VOXEL = 0
-  PHYSICAL = 1
 
 def find_edge_dtype(num_verts:int) -> np.dtype:
   if num_verts < np.dtype(np.uint8).max:
@@ -151,38 +29,32 @@ def find_edge_dtype(num_verts:int) -> np.dtype:
 class OstdHeader:
   MAGIC = b'ostd'
   FORMAT_VERSION = 0
-  HEADER_BYTES = 114
-  # SCHEMA = [
-  #   ['magic', 4 ],
-  #   ['format_version', 1 ],
-  #   ['id', 16 ],
-  #   ['flags', 4 ],
-  #   ['Nv', 8 ], # num vertices
-  #   ['Ne', 8 ], # num edges
-  #   ['Nattr', 2 ],
-  #   ['attr_name_width', 1 ],
-  #   ['num_components', 4 ],
-  #   ['transform', 4*4*4 ],
-  #   ['crc16', 2 ],
-  # ]
+  HEADER_BYTES = 81
 
   def __init__(
     self,
     Nv:int,
     Ne:int,
-    space:Union[str, SpaceType] = SpaceType.VOXEL,
-    vertex_datatype:DataType = DataType.F32,
-    edge_representation:EdgeRepresentation = EdgeRepresentation.PAIR,
-    graph_type:GraphType = GraphType.GRAPH,
-    physical_unit:Union[str, PhysicalUnit] = PhysicalUnit.NANOMETER,
-    compression:CompressionType = CompressionType.NONE,
+    append_mode:bool = False,
+    attribute_header_bytes:int = 0,
     crc16:Optional[int] = None,
-    transform:npt.NDArray[np.float32] = np.eye(4),
-    Nattr:int = 0,
-    attr_name_width:int = 1,
-    num_components:Optional[int] = None,
+    edge_datatype:DataType = DataType.U32,
+    edge_compression:CompressionType = CompressionType.NONE,
+    edge_representation:EdgeRepresentation = EdgeRepresentation.PAIR,
+    edge_bytes:int = 0,
     format_version:int = 0,
+    graph_type:GraphType = GraphType.GRAPH,
+    has_transform:bool = True,
     id:Optional[int] = None,
+    num_axes:Literal[2,3] = 3,
+    num_components:Optional[int] = None,
+    physical_unit:Union[str, PhysicalUnit] = PhysicalUnit.NANOMETER,
+    space:Union[str, SpaceType] = SpaceType.VOXEL,
+    spatial_index_bytes:int = 0,
+    total_bytes:int = 0,
+    vertex_compression:CompressionType = CompressionType.NONE,
+    vertex_datatype:DataType = DataType.F32,
+    vertex_bytes:int = 0,
   ):
     self.Nv = int(Nv)
     self.Ne = int(Ne)
@@ -190,8 +62,10 @@ class OstdHeader:
     if isinstance(space, str) and space == "physical":
       self.space = SpaceType.PHYSICAL
 
+    self.append_mode = append_mode
+
     self.vertex_datatype = vertex_datatype
-    self.edge_dtype = find_edge_dtype(Nv)
+    self.edge_dtype = edge_dtype
     self.edge_representation = edge_representation
     self.graph_type = graph_type
 
@@ -228,18 +102,11 @@ class OstdHeader:
 
   @property
   def edge_dtype(self):
-    if self.Nv < np.dtype(np.uint8).max:
-      return np.uint8
-    elif self.Nv < np.dtype(np.uint16).max:
-      return np.uint16
-    elif self.Nv < np.dtype(np.uint32).max:
-      return np.uint32
-    else:
-      return np.uint64
+    return np.dtype(FROM_DATATYPE[self.edge_datatype])
 
   @property
   def Nvb(self):
-    return self.Nv * 3 * self.vertex_dtype.itemsize
+    return self.Nv * self.num_axes * self.vertex_dtype.itemsize
 
   def encode_flags(self) -> int:
     flags = np.uint32(0)
